--- src/pages/EarthEngineImport/util/isAggregation.js
@@ -1,19 +1,0 @@
-const AGGREGATION_TYPE_SUM = 'sum'
-const AGGREGATION_TYPE_MIN = 'min'
-const AGGREGATION_TYPE_MAX = 'max'
-const AGGREGATION_TYPE_MEAN = 'mean'
-const AGGREGATION_TYPE_MEDIAN = 'median'
-const AGGREGATION_TYPE_STDDEV = 'stdDev'
-const AGGREGATION_TYPE_VARIANCE = 'variance'
-
-const ALL_AGGREGATION_TYPES = [
-    AGGREGATION_TYPE_SUM,
-    AGGREGATION_TYPE_MIN,
-    AGGREGATION_TYPE_MAX,
-    AGGREGATION_TYPE_MEAN,
-    AGGREGATION_TYPE_MEDIAN,
-    AGGREGATION_TYPE_STDDEV,
-    AGGREGATION_TYPE_VARIANCE,
-]
-
-export const isAggregation = (bandId) => ALL_AGGREGATION_TYPES.includes(bandId)
